pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pypsa-nza-plotter"
version = "0.1.0"
description = "Configuration-driven, reproducible plotting for publication-quality scientific figures"
readme = "README.md"
license = {text = "MIT"}
authors = [{name="Your Name", email="you@example.com"}]
requires-python = ">=3.10"

dependencies = [
  "matplotlib>=3.5.0",
  "numpy>=1.21.0",
  "pandas>=1.3.0",
  "PyYAML>=6.0"
]

[tool.setuptools.packages.find]
where = ["src"]
